requirements.txt
tensorflow==2.13.0
flask==3.0.0
numpy==1.24.3
Pillow==10.0.1
opencv-python==4.8.1.78
matplotlib==3.7.2
scikit-learn==1.3.0
gdown==4.7.1
tqdm==4.66.1
